requirements.txt
openai
line-bot-sdk
flask
pyquery
gunicorn